Fixed #1. Changed property type of IHttpRequest and IHttpResponse from int to long.

+++ AbaSoft.Net/IHttpRequest.cs
@@ -13,7 +13,7 @@ namespace AbaSoft.Net
 
         Stream InputStream { get; }
 
-        int ContentLength64 { get; }
+        long ContentLength64 { get; }
 
         NameValueCollection Headers { get; }
